Move QueryHelper to separate module
